RecipeBox page changes
