+++ front-end/src/RecipeBoxPage.js
@@ -6,11 +6,11 @@ import { useEffect, useState } from 'react'
 import axios from 'axios'
 import ComboBoxSearchBar from './ComboBoxSearchBar'
 import Dropdown from 'react-bootstrap/Dropdown'
-import InputGroup from 'react-bootstrap/InputGroup'
 import Button from 'react-bootstrap/Button'
 import ButtonGroup from 'react-bootstrap/ButtonGroup'
-import { ArrowDown, SortUp } from 'react-bootstrap-icons'
+import { SortUp } from 'react-bootstrap-icons'
 import { SortDown } from 'react-bootstrap-icons'
+import TagButton from './TagButton.js'
 
 // Pulls recipes from the database for the logged-in feed and generates a recipelist for qualifying recipes
 // Qualifying recipes, e.g. latest recipes posted by those someone has followed
@@ -61,11 +61,6 @@ const RecipeBoxPage = (props) => {
     // Recipe list to show -- Is later re-assigned if/when sorts are applied
     const [recBoxRecipes, setRecBoxRecipes] = useState([])
 
-    // Sorting -- Expects to be a constantly edited string
-    const [sortBy, setSortBy] = useState('')
-    // Filtering -- Expects to be an array of tags
-    const [filterTags, setFilterTags] = useState([])
-
     // Adjust dropdown and ordering
     const [sortByString, setSortByString] = useState('Sort by Date Posted')
     const [ascendingOrder, setAscendingOrder] = useState(false)
@@ -113,6 +108,59 @@ const RecipeBoxPage = (props) => {
 
     }
 
+    /* Tags -- See BrowseRecipesPage.js */
+    // request all tags on initial render
+    const [tags, setTags] = useState([])
+    const [filterKeyword, setFilterKeyword] = useState('')
+    const [filterTags, setFilterTags] = useState([])
+
+    useEffect(() => {
+        // fetch all tags
+        axios('http://localhost:4000/tags')
+        .then((response) => {
+            setTags(response.data)
+        })
+        .catch((err) => {
+            console.error(err)
+
+            // make some backup fake data
+            const backupData = [
+                {
+                    tag: 'vegan',
+                    count: 10,
+                    id: 1
+                },
+                {
+                    tag: 'appetizer',
+                    count: 34,
+                    id: 2
+                },
+                {
+                    tag: 'mexican',
+                    count: 22,
+                    id: 3
+                }
+            ]
+
+            setTags(backupData.map(tag => tag.tag))
+        })
+    }, [])
+    
+    const [tagSelection, setTagSelection] = useState('')
+
+    useEffect(() => {
+        if (tagSelection !== '') {
+            // append selected tag to filter tags array
+            setFilterTags(filterTags.concat([tagSelection]))
+
+            // remove selected tag from tags array
+            const tagIndex = tags.indexOf(tagSelection)
+            setTags(tags.slice(0, tagIndex).concat(tags.slice(tagIndex + 1)))
+        }
+
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+    }, [tagSelection])
+
     return (
         <div className="container">
             {/* Title */}
@@ -137,7 +185,13 @@ const RecipeBoxPage = (props) => {
                   
                 </div>
 
-                <ComboBoxSearchBar isTag={true} tags={filterTags} setSelection={setFilterTags} />
+                <div className="recipeNameSearchbar">
+                    <KeywordSearchBar isRecipe={true} filter={filterKeyword} setFilter={setFilterKeyword}/>
+                </div>
+                <ComboBoxSearchBar isTag={true} tags={tags} setSelection={setTagSelection} />
+                <div className="tagButtonsSection">
+                    {filterTags.map((tag, i) => <TagButton tag={tag} filterTags={filterTags} setFilterTags={setFilterTags} tags={tags} setTags={setTags} key={i} />)}
+                </div>
             </div>
 
             {/* Generate the list of recipes */}
